# backend/openapi/epa-platform-v1.yaml
openapi: 3.1.0
info:
  title: EPA Platform API
  version: 1.0.0
  description: |
    HIPAA-compliant electronic Prior Authorization (ePA) platform API aligned with
    HL7 FHIR R4 and SMART on FHIR v2.0.1.

    ## Audit & Immutability

    All mutating operations are **append-only**. Prior authorization submissions and
    eligibility checks cannot be deleted or overwritten; corrections are submitted as
    new resources linked via `Claim.related` or `CoverageEligibilityRequest.basedOn`.

    Every request MUST include audit correlation headers. Responses include
    `X-Audit-Event-Id` and `X-Resource-Version` for immutable audit trail linkage.

    ## FHIR Content

    Request and response bodies use FHIR R4 JSON (`application/fhir+json`) unless
    noted. See `fhir-profiles/README.md` for StructureDefinition mappings.

  contact:
    name: EPA Platform API Support
    email: api-support@epa-platform.example.com
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
...
servers:
  - url: https://api.epa-platform.example.com
    description: Production
  - url: https://api.staging.epa-platform.example.com
    description: Staging
  - url: http://localhost:8000
    description: Local development
...
security:
  - smartOAuth:
      - openid
      - fhirUser
      - launch
      - offline_access
paths:
...
  /oauth/authorize:
    get:
      tags: [OAuth]
      operationId: oauthAuthorize
      summary: OAuth 2.0 authorization endpoint (SMART App Launch)
      description: |
        Initiates the SMART authorization code flow with PKCE (S256 required).
        Supports `launch` (EHR) and standalone patient/provider flows.
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: true
          description: |
            Space-delimited SMART scopes. Example:
            `launch patient/Patient.read patient/Coverage.read user/Claim.write openid fhirUser`
          schema:
            type: string
            example: launch patient/Patient.read user/Claim.write openid fhirUser
        - name: state
          in: query
          required: true
          schema:
            type: string
            minLength: 8
        - name: aud
          in: query
          required: true
          description: FHIR server base URL (resource server audience)
          schema:
            type: string
            format: uri
            example: https://api.epa-platform.example.com/fhir/R4
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
        - name: launch
          in: query
          required: false
          description: EHR launch context handle (required for EHR launch)
          schema:
            type: string
        - name: nonce
          in: query
          required: false
          description: Required when `openid` scope is requested
          schema:
            type: string
      responses:
        '302':
          description: Redirect to client redirect_uri with authorization code
          headers:
            Location:
              schema:
                type: string
                format: uri
        '400':
          $ref: '#/components/responses/OAuthError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /oauth/token:
    post:
      tags: [OAuth]
      operationId: oauthToken
      summary: OAuth 2.0 token endpoint
      description: |
        Exchanges authorization codes or refresh tokens for access tokens.
        Supports `authorization_code` and `refresh_token` grant types.
        Client authentication via `client_secret_basic`, `client_secret_post`,
        or `private_key_jwt` (preferred for confidential apps).
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              oneOf:
                - $ref: '#/components/schemas/AuthorizationCodeTokenRequest'
                - $ref: '#/components/schemas/RefreshTokenRequest'
                - $ref: '#/components/schemas/ClientCredentialsTokenRequest'
            examples:
              authorization_code:
                summary: Authorization code exchange with PKCE
                value:
                  grant_type: authorization_code
                  code: AUTH_CODE_EXAMPLE
                  redirect_uri: https://app.example.com/callback
                  client_id: epa-client-001
                  code_verifier: dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk
              refresh_token:
                summary: Refresh token rotation
                value:
                  grant_type: refresh_token
                  refresh_token: REFRESH_TOKEN_EXAMPLE
                  client_id: epa-client-001
                  scope: patient/Patient.read user/Claim.write openid
      responses:
        '200':
          description: Token response
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
            Pragma:
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/OAuthError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /oauth/introspect:
    post:
      tags: [OAuth]
      operationId: oauthIntrospect
      summary: OAuth 2.0 token introspection (RFC 7662)
      description: |
        Validates access or refresh tokens. Restricted to confidential resource
        servers and authorized audit services. Returns active status and granted scopes.
      security:
        - clientCredentials: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Access or refresh token to introspect
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
      responses:
        '200':
          description: Token introspection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenIntrospectionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /oauth/revoke:
    post:
      tags: [OAuth]
      operationId: oauthRevoke
      summary: OAuth 2.0 token revocation (RFC 7009)
      security:
        - clientCredentials: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
      responses:
        '200':
          description: Token revoked (always returns 200 per RFC 7009)
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    smartOAuth:
      type: oauth2
      description: |
        SMART on FHIR OAuth 2.0 with PKCE. Scopes follow the pattern
        `{context}/{ResourceType}.{permission}` where context is `patient`, `user`,
        or `system`.
      flows:
        authorizationCode:
          authorizationUrl: https://api.epa-platform.example.com/oauth/authorize
          tokenUrl: https://api.epa-platform.example.com/oauth/token
          refreshUrl: https://api.epa-platform.example.com/oauth/token
          scopes:
            openid: OpenID Connect identity
            fhirUser: FHIR Practitioner/Patient/User resource URL for current user
            launch: EHR launch context
            launch/patient: Request patient context at launch
            offline_access: Issue refresh tokens
            online_access: Issue access tokens (implicit in code flow)
            patient/Patient.read: Read patient demographics (patient context)
            patient/Coverage.read: Read patient coverage (patient context)
            patient/Claim.read: Read patient prior auth claims (patient context)
            patient/Claim.write: Submit patient prior auth claims (patient context)
            patient/CoverageEligibilityRequest.write: Submit eligibility checks (patient context)
            user/Patient.read: Read any patient the user has access to
            user/Coverage.read: Read coverage for accessible patients
            user/Claim.read: Read prior auth claims
            user/Claim.write: Submit prior authorization requests
            user/CoverageEligibilityRequest.write: Submit coverage eligibility requests
            user/CoverageEligibilityResponse.read: Read eligibility responses
            system/Claim.read: Backend service read all claims
            system/AuditEvent.read: Audit service read audit events

    clientCredentials:
      type: oauth2
      description: Client credentials for confidential server-to-server clients (introspection, audit)
      flows:
        clientCredentials:
          tokenUrl: https://api.epa-platform.example.com/oauth/token
          scopes:
            system/AuditEvent.read: Read audit trail
...
  responses:
...
    Unauthorized:
      description: Missing or invalid access token
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer error="invalid_token", error_description="Token expired"
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    OAuthError:
      description: OAuth 2.0 error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OAuthErrorResponse'
  schemas:
...
    AuthorizationCodeTokenRequest:
      type: object
      required: [grant_type, code, redirect_uri, client_id, code_verifier]
      properties:
        grant_type:
          type: string
          const: authorization_code
        code:
          type: string
        redirect_uri:
          type: string
          format: uri
        client_id:
          type: string
        client_secret:
          type: string
          description: Required for confidential clients not using private_key_jwt
        code_verifier:
          type: string
        client_assertion_type:
          type: string
          enum:
            - urn:ietf:params:oauth:client-assertion-type:jwt-bearer
        client_assertion:
          type: string
          description: JWT signed by client for private_key_jwt authentication
    RefreshTokenRequest:
      type: object
      required: [grant_type, refresh_token]
      properties:
        grant_type:
          type: string
          const: refresh_token
        refresh_token:
          type: string
        client_id:
          type: string
        scope:
          type: string
          description: Optional scope reduction
    ClientCredentialsTokenRequest:
      type: object
      required: [grant_type]
      properties:
        grant_type:
          type: string
          const: client_credentials
        scope:
          type: string
        client_id:
          type: string
        client_secret:
          type: string
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 3600
        scope:
          type: string
          description: Granted scopes (space-delimited)
        refresh_token:
          type: string
        patient:
          type: string
          description: FHIR Patient resource id (patient-scoped tokens)
        encounter:
          type: string
          description: FHIR Encounter resource id when in encounter context
        id_token:
          type: string
          description: OpenID Connect ID token (when openid scope granted)
        need_patient_banner:
          type: boolean
        smart_style_url:
          type: string
          format: uri
    TokenIntrospectionResponse:
      type: object
      required: [active]
      properties:
        active:
          type: boolean
        scope:
          type: string
        client_id:
          type: string
        username:
          type: string
        token_type:
          type: string
        exp:
          type: integer
          description: Expiration time (Unix epoch)
        iat:
          type: integer
        sub:
          type: string
          description: Subject (often fhirUser reference)
        aud:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        iss:
          type: string
        jti:
          type: string
        patient:
          type: string
          description: Patient context for patient-scoped tokens
        fhirUser:
          type: string
          format: uri
          description: FHIR resource URL of the authenticated user
    OAuthErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - access_denied
        error_description:
          type: string
        error_uri:
          type: string
          format: uri
    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: https://api.epa-platform.example.com/problems/validation-error
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        requestId:
          type: string
          format: uuid
        auditEventId:
          type: string
    OperationOutcome:
      type: object
      description: FHIR R4 OperationOutcome for error and warning reporting
      required: [resourceType, issue]
      properties:
        resourceType:
          type: string
          const: OperationOutcome
        id:
          type: string
        issue:
          type: array
          minItems: 1
          items:
            type: object
            required: [severity, code]
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
                description: IssueType code from http://hl7.org/fhir/issue-type
              details:
                type: object
                properties:
                  text:
                    type: string
              diagnostics:
                type: string
              expression:
                type: array
                items:
                  type: string
...